% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-datafile.R
\name{add_datafile}
\alias{add_datafile}
\title{Ajoute un datafile à un dataset}
\usage{
add_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = format(Sys.time(), "\%Y-\%m"),
  published = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S")
)
}
\arguments{
\item{dataset}{l'id d'un dataset, un objet \code{dido_dataset()},
\code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}

\item{published}{la date/heure de publication du fichier, si non précisée, prend la
date/heure du moment.

Ce paramètre est au format "AAAA-MM-JJ HH:MM:SS" (ou ISO8601 si vous
préférez). Si la timezone n'est pas précisée, la timezone de l'ordinateur
local est utilisée.}

\item{temporal_coverage_start}{optionnel, la date de début de couverture du
fichier de données au format AAAA-MM-JJ}

\item{temporal_coverage_end}{optionnel, la date de fin de couverture du
fichier de données au format AAAA-MM-JJ}

\item{legal_notice}{les mentions légales, par défaut "SDES"}

\item{date_diffusion}{(optionnel) les date et heure auxquelles le fichier
devra être rendu accessible à la diffusion.

Ce paramètre est au format "AAAA-MM-JJ HH:MM:SS" (ou ISO8601 si vous
préférez). Si la timezone n'est pas précisée, la timezone de l'ordinateur
local est utilisée.

Si cette date/heure est dans le passé, les données sont immédiatement
accessibles, si elle est dans le futur, les données ne seront accessibles
qu'à cette date/heure.

Si ce paramètre est non précisé prend la date/heure courante, les données
sont donc immédiatement accessibles.}
}
\value{
un objet \code{dido_job()}
}
\description{
Ajoute un datafile à un dataset
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dataset <- add_or_update_dataset(
  title = "Un dataset pour les datafiles",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "unknown",
)

add_datafile(
  dataset = dataset,
  title = "titre",
  description = "description",
  file_name = dido_example("augmente.csv")
)

# publier un fichier de données avec toutes les métadonnées et un embargo
add_datafile(
  dataset = dataset,
  title = "titre 2",
  description = "description 2",
  file_name = dido_example("augmente.csv"),
  temporal_coverage_start = "2021-01-01",
  temporal_coverage_end = "2021-12-31",
  legal_notice = "something",
  millesime = "2020-10",
  date_diffusion = "2020-11-01 07:45:00"
)
}
\seealso{
Other datafile: 
\code{\link{add_or_update_datafile}()},
\code{\link{delete_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
