% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-millesime.R
\name{add_millesime}
\alias{add_millesime}
\title{Ajoute un millésime à un datafile}
\usage{
add_millesime(
  datafile,
  file_name,
  date_diffusion = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  millesime = format(Sys.time(), "\%Y-\%m"),
  quiet = NULL
)
}
\arguments{
\item{datafile}{un objet dido_datafile obtenu par \code{get_datafile()}}

\item{file_name}{le nom du fichier à charger}

\item{date_diffusion}{(optionnel) la date/heure à laquelle le fichier
devra être rendu accessible à la diffusion au format "AAAA-MM-JJ HH:MM:SS".
Si la time zone n'est pas précisée, l'heure locale est utilisée.

Si cette date est dans le passé, les données sont immédiatement
accessibles, si elle est dans le futur, les données ne seront accessibles
qu'à cette date/heure.

Si non précisée prend la date/heure courante (heure locale), les données
sont donc immédiatement accessibles.}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}

\item{quiet}{quand \code{TRUE} ou que l'option dido_quiet est à \code{TRUE} supprime
les messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{dido_job()}
}
\description{
Ajoute un millésime à un datafile
}
\examples{
library(dplyr, warn.conflicts = FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test")

millesime <- add_millesime(
  datafile = datafile,
  file_name = dido_example("augmente.csv"),
  millesime = "2011-10"
)


# publier un millésime avec embargo
millesime <- add_millesime(
  datafile = datafile,
  file_name = dido_example("augmente.csv"),
  millesime = "2012-10",
  date_diffusion = "2011-10-10 08:00:00"
)
}
\seealso{
Other millesime: 
\code{\link{list_millesimes}()},
\code{\link{replace_millesime}()}
}
\concept{millesime}
