% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-or-update-datafile.R
\name{add_or_update_datafile}
\alias{add_or_update_datafile}
\title{Ajoute ou modifie un datafile}
\usage{
add_or_update_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = format(Sys.time(), "\%Y-\%m"),
  published = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  keep_old_millesimes = Inf,
  on_existing_millesime = "fail",
  check_file_date = FALSE,
  quiet = NULL
)
}
\arguments{
\item{dataset}{l'id d'un dataset, un objet \code{dido_dataset()},
\code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}

\item{published}{la date/heure de publication du fichier, si non précisée, prend la
date/heure du moment.

Ce paramètre est au format "AAAA-MM-JJ HH:MM:SS" (ou ISO8601 si vous
préférez). Si la timezone n'est pas précisée, la timezone de l'ordinateur
local est utilisée.}

\item{temporal_coverage_start}{optionnel, la date de début de couverture du
fichier de données au format AAAA-MM-JJ}

\item{temporal_coverage_end}{optionnel, la date de fin de couverture du
fichier de données au format AAAA-MM-JJ}

\item{legal_notice}{les mentions légales, par défaut "SDES"}

\item{date_diffusion}{(optionnel) les date et heure auxquelles le fichier
devra être rendu accessible à la diffusion.

Ce paramètre est au format "AAAA-MM-JJ HH:MM:SS" (ou ISO8601 si vous
préférez). Si la timezone n'est pas précisée, la timezone de l'ordinateur
local est utilisée.

Si cette date/heure est dans le passé, les données sont immédiatement
accessibles, si elle est dans le futur, les données ne seront accessibles
qu'à cette date/heure.

Si ce paramètre est non précisé prend la date/heure courante, les données
sont donc immédiatement accessibles.}

\item{keep_old_millesimes}{nombre d'ancien millésimes à conserver. Ce
paramètre permet de supprimer les anciens millésimes. Pour des raisons de
sécurité, cette option n'effacera jamais des millésimes plus récent que le
millésime passé en paramètre.}

\item{on_existing_millesime}{skip/fail/replace : action à faire quand le
millésime existe déjà. Peut-être :
\itemize{
\item "skip" : retourne immédiatement après avoir affiché un message. Les métadonnées du millésime ne sont pas mises à jour et les anciens millésimes ne sont jamais effacés.
\item "fail" : lève une exception de class \code{existing_millesime}. Les métadonnées du millésime ne sont pas mises à jour et les anciens millésimes ne sont jamais effacés.
\item "replace" : remplace le millésime ayant le même identifiant
}}

\item{check_file_date}{TRUE/FALSE, Si TRUE met à jour le datafile uniquement
si le fichier est plus récent que le last_modified du datafile}
}
\value{
un objet \code{dido_job()} ou \code{NULL} si aucune création/mise à jour
n'a eu lieu
}
\description{
met à jour le datafile (métadonnées et données) avec le même titre s'il
existe sinon ajoute un datafile, un millésime ou remplace un millésime s'il
existe déjà.

Le paramètre \code{on_existing_millesime} permet de préciser le comportement à
adopter si le millésime existe déjà.

Si plusieurs datafiles ont le même titre, cette méthode retourne une erreur.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dataset <- add_or_update_dataset(
  title = "Un dataset pour les add_or_update_datafiles",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "unknown",
)

add_or_update_datafile(
  dataset = dataset,
  title = "titre",
  description = "description",
  file_name = dido_example("augmente.csv"),
  millesime = "2021-10",
)

add_or_update_datafile(
  dataset = dataset,
  title = "titre",
  description = "description",
  file_name = dido_example("augmente.csv"),
  temporal_coverage_start = "2021-01-01",
  temporal_coverage_end = "2021-12-31",
  millesime = "2022-10",
)
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{delete_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
