% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-or-update-dataset.R
\name{add_or_update_dataset}
\alias{add_or_update_dataset}
\title{Ajoute ou met à jour un dataset}
\usage{
add_or_update_dataset(
  title,
  description,
  topic,
  frequency,
  tags = NULL,
  frequency_date = NULL,
  granularity = NULL,
  zones = NULL,
  organization = my_organization(),
  license = "fr-lo",
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  caution = NULL,
  quiet = NULL
)
}
\arguments{
\item{title}{le titre du dataset}

\item{description}{la description du dataset}

\item{topic}{le thème du dataset. Doit être dans "Environnement", "Énergie",
"Transports", "Logement", "Changement climatique"}

\item{frequency}{la fréquence de publication des données. Les fréquences les
plus utilisées sont \code{annual}, \code{monthly}, \code{ponctual}, \code{unknown}. \href{https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/metadata/frequencies}{la liste complète des fréquences}}

\item{tags}{la liste des mots clefs. Si le fichier est déjà publié sur le
site SDES, on reprend les mêmes. La liste des \href{https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/referentiels/tags/csv?withColumnName=true&withColumnDescription=false}{mots clefs recensés}.}

\item{frequency_date}{la date de prochaine publication. optionnel seulement
la \code{frequency} est \code{unknown}.}

\item{granularity}{la granularité du jeu de données. Les plus utilisés sont
"fr:region", "fr:departement", "fr:epci", "fr:commune", "fr:iris" et pour
les données à l'adresse "poi". \href{https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/metadata/granularities}{La liste complète des granularités}}

\item{zones}{la zone couverte par le jeu de données. \code{country:fr} pour France
entière, \code{country-subset:fr:metro} pour la métropole et
`country-subset:fr:drom`` pour les DROM}

\item{organization}{l'id de l'organisation sous laquelle vous souhaitez
publier le dataset, si vous n'appartenez qu'à une seule organisation, vous
n'avez pas à remplir ce champ, didoscalim la prendra par défaut.

Si ça n'est pas le cas, vous devez préciser la fonction \code{my_organization()}}

\item{license}{la licence des données. "fr-lo" par défaut}

\item{temporal_coverage_start}{optionnel, la date de début de couverture du
jeux de données au format AAAA-MM-JJ}

\item{temporal_coverage_end}{optionnel, la date de fin de couverture du jeux
de données au format AAAA-MM-JJ}

\item{caution}{Les précautions à prendre avec ce jeu de données. Par exemple
: "certaines données peuvent être secrétisées"}
}
\value{
un objet \code{\link[=dido_dataset]{dido_dataset()}}. Ce dernier sert pour créer par la suite
les datafiles, les millésimes et les attachments
}
\description{
Cette fonction permet, par exemple, de modifier la période de couverture du
dataset ainsi que la date de prochaine mise à jour.
}
\examples{
dataset <- add_or_update_dataset(
  title = "Des données statistiques",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "annual",
  frequency_date = "2021-01-10",
  temporal_coverage_start = "2020-01-01",
  temporal_coverage_end = "2020-12-31"
)

dataset <- add_or_update_dataset(
  title = "Des données statistiques",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "annual",
  frequency_date = "2022-01-10",
  temporal_coverage_start = "2021-01-01",
  temporal_coverage_end = "2021-12-31"
)
}
\seealso{
La liste des \href{https://data.statistiques.developpement-durable.gouv.fr/dido/api/v1/referentiels/tags/csv?withColumnName=true&withColumnDescription=false}{tags/mots clefs recensés}

Other dataset: 
\code{\link{add_dataset}()},
\code{\link{dido_dataset}()},
\code{\link{get_dataset}()},
\code{\link{update_dataset}()}
}
\concept{dataset}
