% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{check_config}
\alias{check_config}
\title{Vérifier l'accès aux différents environnements}
\usage{
check_config()
}
\description{
Affiche un état des connexions aux différents environnements avec le message
d'erreur le cas échéant
}
\examples{
\dontrun{
check_config()
}
}
