% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{check_envs}
\alias{check_envs}
\title{Vérifier l'accès aux différents environnements}
\usage{
check_envs()
}
\description{
Affiche un état des connexions aux différents environnements avec le message
d'erreur le cas échéant
}
\examples{
\dontrun{
check_envs()
}
}
