% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-import-file.R
\name{check_import_file}
\alias{check_import_file}
\title{Vérifie un fichier CSV}
\usage{
check_import_file(file_name)
}
\arguments{
\item{file_name}{le fichier a valider}
}
\value{
TRUE si l'import s'est bien passé, FALSE sinon
}
\description{
Cette fonction permet de vérifier rapidement la validité d'un fichier de
données. Elle :
\enumerate{
\item créé un jeu de données avec un nom aléatoire
\item ajoute le fichier de données
\item supprime le jeu de données
\item affiche les erreurs rencontrées s'il y en a
}

\strong{L'import de test ne se fera jamais sur l'environnement de production} mais
sur un des environnments DEV, PREPROD ou ECOLE. Si aucun de ces
environnements n'est configuré, la fonction remonte une erreur
}
\examples{

check_import_file(dido_example("file-with-error.csv"))
}
