% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R
\name{clean_datafile_metadata}
\alias{clean_datafile_metadata}
\title{Nettoie l'objet retourné par \code{get_datafile()}}
\usage{
clean_datafile_metadata(df)
}
\arguments{
\item{df}{l'objet retourné par \code{get_datafile()}}
}
\value{
un objet \code{datafile} que l'on peut passer à \code{update_datafile()}
}
\description{
Nettoie l'objet retourné par \code{get_datafile()}
}
\examples{
\dontrun{
df <- get_datafile("id", "rid")
df$temporal_coverage$end <- "2023-12-31"
update_datafile("id", "rid", df)
}
}
