% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Prépare un objet pour l'envoyer à DiDo}
\usage{
clean_metadata(data)
}
\arguments{
\item{data}{objet dido_dataset, dido_datafile}
}
\value{
un objet nettoyé
}
\description{
Prépare un objet pour l'envoyer à DiDo
}
\examples{
\dontrun{
clean_metadata(dataset)
}
}
\keyword{internal}
