% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-datafile.R
\name{create_datafile}
\alias{create_datafile}
\title{Créé un datafile}
\usage{
create_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = NULL,
  published = NULL,
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{l'id d'un dataset ou un objet dataset tel que retourné par
\code{get_dataset()} ou \code{create_dataset()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (YYYY-MM). Par défaut prendre la valeur YYYY-MM
avec l'année courante et le mois courant}

\item{published}{la date de publication du fichier, si non précisée, prend la
date du jour.}

\item{temporal_coverage_start}{la date de début des données au format
YYYY-MM-DD}

\item{temporal_coverage_end}{la date de fin des données au format YYYY-MM-DD}

\item{legal_notice}{mention légale, par défaut "SDES"}

\item{date_diffusion}{date/heure de diffusion du fichier au format ISO 8601
(2021-10-01T08:00:00Z). Si non précisée prend la date du jour à minuit.}

\item{quiet}{Si TRUE n'affiche pas les messages. Défaut à \code{FALSE}}
}
\value{
un objet \code{\link[=job_result]{job_result()}}
}
\description{
Créé un datafile
}
\examples{
\dontrun{
datafile <- create_datafile(
  dataset = dataset$id,
  title = "titre",
  description = "description",
  token_file = file_id
)
}
}
