% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_heure_iso8601}
\alias{date_heure_iso8601}
\title{Retourne une date au format ISO 8601 avec la TZ}
\usage{
date_heure_iso8601(date_heure)
}
\arguments{
\item{date_heure}{date/heure}
}
\value{
une date/heure au format ISO  8601 avec la TZ
}
\description{
Retourne une date au format ISO 8601 avec la TZ
}
\examples{
date_heure_iso8601("2022-02-02 07:00:02")
}
\keyword{internal}
