% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-attachment.R
\name{delete_attachment}
\alias{delete_attachment}
\title{Supprime un attachment}
\usage{
delete_attachment(attachment)
}
\arguments{
\item{attachment}{un objet attachment retourné par \code{get_attachment()}, \code{add_attachment()}}
}
\value{
TRUE
}
\description{
Supprime un attachment
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dataset <- add_or_update_dataset(
  title = "Un dataset pour les attachements",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "unknown",
)

attachment <- add_attachment(
  dataset = dataset,
  title = "title",
  description = "description",
  file_name = dido_example("attachment.txt")
)
delete_attachment(attachment)
}
\seealso{
Other attachment: 
\code{\link{add_attachment}()},
\code{\link{add_or_update_attachment}()},
\code{\link{dido_attachment}()},
\code{\link{get_attachment}()},
\code{\link{list_attachments}()},
\code{\link{replace_attachment}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
