% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-datafile.R
\name{delete_datafile}
\alias{delete_datafile}
\title{Supprime un datafile.}
\usage{
delete_datafile(datafile)
}
\arguments{
\item{datafile}{un objet datafile retourné par \code{get_datafile()}, \code{add_datafile()}}
}
\value{
TRUE
}
\description{
Supprime un datafile.
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{add_or_update_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
