% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Supprime un dataset}
\usage{
delete_dataset(dataset)
}
\arguments{
\item{dataset}{un identifiant de dataset}
}
\value{
TRUE
}
\description{
Supprime un dataset
}
\examples{
dataset <- add_dataset(
  title = "le titre du dataset",
  description = "la description du dataset",
  topic = "Transports",
  frequency = "unknown"
)
delete_dataset(dataset)
}
\keyword{internal}
