% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-millesime.R
\name{delete_millesime}
\alias{delete_millesime}
\title{Supprimer un millésime}
\usage{
delete_millesime(datafile, millesime)
}
\arguments{
\item{datafile}{un objet dido_datafile obtenu par \code{get_datafile()}}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}
}
\value{
TRUE
}
\description{
Supprimer un millésime
}
\seealso{
Other millesime: 
\code{\link{add_millesime}()},
\code{\link{list_millesimes}()},
\code{\link{replace_millesime}()}
}
\concept{millesime}
