% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-api.R
\name{dido_api}
\alias{dido_api}
\title{Envoie une requête au serveur DiDo}
\usage{
dido_api(
  method,
  path,
  body,
  query_params = list(),
  headers = c(),
  as_tibble = FALSE
)
}
\arguments{
\item{method}{une des méthodes GET/POST/PUT/DELETE}

\item{path}{le chemin de l'api}

\item{body}{le body de la requête}

\item{query_params}{les paramètres de la requête}

\item{headers}{les entêtes de la requête}

\item{as_tibble}{TRUE/FALSE si TRUE retourne un tibble à la place d'un objet.
Défaut à FALSE}
}
\value{
un objet json ou un dataframe
}
\description{
Envoie la requête au serveur et retourne le résultat sous forme d'un objet ou un dataframe si as_dataframe est à TRUE
}
\examples{
\dontrun{
job <- dido_api(method = "POST", path = url, body = body)
alerts <- dido_api(method = "GET", path = "/datasets/alerts", as_tibble = TRUE)
}
}
\keyword{internal}
