% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-attachment.R
\name{dido_attachment}
\alias{dido_attachment}
\title{L'objet dido_attachment}
\description{
L'objet dido_attachment inclus les champs :
The response object captures all information from a request.  It includes
fields:
}
