% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-datafile.R
\name{dido_datafile}
\alias{dido_datafile}
\title{L'objet dido_datafile}
\description{
L'objet dido_datafile et retourné par \code{get_datafile()} et peut-être utilisé
en paramètre par les fonctions de manipulation de dataset, datafile,
millésime et attachement.

Il a les champs suivants :
\itemize{
\item \code{rid} l'identifiant du datafile
\item \code{title}
\item \code{description}
\item \code{millesimes} le nb de millésimes dans le datafile
\item \code{created_at}
\item \code{last_modified}
\item \code{published}
\item \code{url}
\item \code{temporal_coverage} un liste avec deux champs \code{start} et \code{end}
\item \code{legal_notice} la licence s'appliquant au fichier de données
\item \code{millesimes_info} comprenant une liste des millésimes rattachés
}

le dernier champ est supprimé par \code{clean_metadata()}

L'objet a également un attribut \code{id} avec l'identifiant du dataset auquel il
appartient
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
