% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-dataset.R
\name{dido_dataset}
\alias{dido_dataset}
\title{L'objet dido_dataset}
\description{
L'objet dido_dataset est retourné par \code{get_dataset()} et \code{add_dataset()}
et utilisé en paramètre par \code{update_dataset()} \code{add_datafile()} et
\code{create_attachment()}
}
