% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{dido_example}
\alias{dido_example}
\title{Retourne le chemin des fichiers exemple}
\usage{
dido_example(file = NULL)
}
\arguments{
\item{file}{le nom du fichier. Si \code{NULL} les exemples sont listés.}
}
\description{
Les fichiers exemples sont stockés dans \code{inst/extdata}
}
\examples{
dido_example()
dido_example("exemple.csv")
#
}
