% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-read-delim.R
\name{dido_read_delim}
\alias{dido_read_delim}
\title{Lit un fichier CSV}
\usage{
dido_read_delim(file, delim = ";", ...)
}
\arguments{
\item{file}{Either a path to a file, a connection, or literal data
(either a single string or a raw vector).

Files ending in \code{.gz}, \code{.bz2}, \code{.xz}, or \code{.zip} will
be automatically uncompressed. Files starting with \verb{http://},
\verb{https://}, \verb{ftp://}, or \verb{ftps://} will be automatically
downloaded. Remote gz files can also be automatically downloaded and
decompressed.

Literal data is most useful for examples and tests. To be recognised as
literal data, the input must be either wrapped with \code{I()}, be a string
containing at least one new line, or be a vector containing at least one
string with a new line.

Using a value of \code{\link[readr:clipboard]{clipboard()}} will read from the system clipboard.}

\item{delim}{le séparateur de champ. Par défaut ";".}

\item{...}{ces arguments seront passés directement à la fonction
\code{readr::read_delim()}}

\item{locale}{la locale à utiliser pour lire les fichiers. A ce niveau, le
paramètre le plus important est l'encodage du fichier. Le défaut est
"UTF-8". Les autres encodages fréquemment rencontrés sont "WINDOWS-1252" et
"ISO-8859-15"}
}
\value{
un tibble dont toutes les colonnes sont de type \code{chr}
}
\description{
Cette fonction utilise directement \code{readr::read_delim} en enlevant la
détection du type des colonnes et en fixant \code{delim} à ";" par défaut.
}
\details{
Certaines variables peuvent avoir des valeurs secrétisées
représentées par la valeur \code{secret}, la détection automatique du package
\code{readr} n'est donc pas fiable et est désactivé à ce niveau. La détection
automatique est faite dans la fonction \code{dido_csv()}.
}
\examples{
dido_read_delim(dido_example("exemple.csv"))
dido_read_delim(dido_example("csv-win-char.csv"),
  delim = ",",
  locale = locale(encoding = "WINDOWS-1252")
)
}
\seealso{
Other csv: 
\code{\link{dido_csv}()},
\code{\link{dido_write_csv}()}
}
\concept{csv}
