% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dido_upload_file}
\alias{dido_upload_file}
\title{Verse un fichier de données à intégrer}
\usage{
dido_upload_file(file_name)
}
\arguments{
\item{file_name}{le nom du fichier}
}
\value{
un token
}
\description{
Verse un fichier de données à intégrer
}
\examples{
token <- dido_upload_file(dido_example("attachment.txt"))
}
\keyword{internal}
