% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-write-csv.R
\name{dido_write_csv}
\alias{dido_write_csv}
\title{Enregistre le fichier CSV augmenté généré}
\usage{
dido_write_csv(data, file)
}
\arguments{
\item{data}{un tibble retourné par \code{dido_csv()}}

\item{file}{le nom du fichier}
}
\value{
le tibble passé en entrée
}
\description{
Enregistre le fichier CSV augmenté généré
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
dido_read_delim(dido_example("exemple.csv")) \%>\%
  dido_csv() \%>\%
  dido_write_csv("/tmp/fichier.csv")
}
