% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didoscalim.R
\docType{package}
\name{didoscalim-package}
\alias{didoscalim-package}
\alias{_PACKAGE}
\title{didoscalim: Charger des données dans DiDo}
\description{
Ce package permet de charger des données dans DiDo en passant par l'API d'alimentation.
}
\author{
\strong{Maintainer}: Nicolas Chuche \email{nicolas.chuche@developpement-durable.gouv.fr}

}
\keyword{internal}
