% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{didoscalim_log}
\alias{didoscalim_log}
\alias{local_didoscalim_verbosity}
\alias{with_didoscalim_verbosity}
\alias{didoscalim_debug}
\alias{didoscalim_info}
\alias{didoscalim_abort}
\title{Journalisation et niveau de verbosité}
\usage{
local_didoscalim_verbosity(level, env = parent.frame())

with_didoscalim_verbosity(level, code)

didoscalim_debug(text, .envir = parent.frame())

didoscalim_info(text, .envir = parent.frame())

didoscalim_abort(
  message = NULL,
  ...,
  class = NULL,
  .envir = parent.frame(),
  call = caller_env()
)
}
\description{
Des méthodes liées à la gestion de la verbosité et à la journalisation
}
\section{\code{local_didoscalim_verbosity}}{

change le niveau de verbosité dans la suite du code au même scope
}

\section{\code{with_didoscalim_verbosity}}{

change le niveau de verbosité dans le bloc de code
}

\section{\code{didoscalim_debug()}}{

affiche les messages de debug
}

\section{\code{didoscalim_info()}}{

affiche les messages d'info
}

\section{\code{didoscalim_abort}}{

affiche le message au niveau le plus elevé et lève une exception.
}

