% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-only.R
\name{didoscalim_update_only}
\alias{didoscalim_update_only}
\title{Bloquer les créations de datasets, datafiles et attachments}
\usage{
didoscalim_update_only(bool)
}
\arguments{
\item{bool}{TRUE/FALSE, si TRUE bloque les créations}
}
\value{
Le précédente valeur ou FALSE si non fixée
}
\description{
\code{didoscalim_update_only()} permet d'autoriser ou d'interdire la création
d'objets dans les fonctions \code{add_or_update_dataset()},
\code{add_or_update_datafile()} et \code{add_or_update_attachment()} dans la suite du
programme
}
\examples{
# bloque la création dans la suite du programme
didoscalim_update_only(TRUE)

# autorise la création dans la suite du programme
didoscalim_update_only(FALSE)
}
