% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ui.R
\name{didoscalim_verbosity}
\alias{didoscalim_verbosity}
\title{Voir et modifier le niveau de message de didoscalim}
\usage{
didoscalim_verbosity()
}
\description{
\code{didoscalim_verbosity()} retourne l'option nommée "didoscalim_verbosity", qui
détermine la verbosité générale de didoscalim. Il y a trois niveaux possibles
:
\itemize{
\item "debug": Niveau pour débugger
\item "info" (défaut): Niveau normal d'information utile à l'utilisateur.
\item "silent": Aucun message sauf les erreurs.
}

Si vous souhaitez changer le niveau par défaut, il suffit d'utiliser \code{option}
:

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(
  didoscalim_verbosity = "silent",
)
}\if{html}{\out{</div>}}
}
\examples{
didoscalim_verbosity()
options(
  didoscalim_verbosity = "silent"
)
didoscalim_verbosity()
}
