% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Retourne un objet avec uniquement ses métadonnées propre}
\usage{
extract_metadata(data)
}
\arguments{
\item{data}{un objet dido_dataset, dido_datafile}
}
\value{
un objet réduit à ses propres métadonnées
}
\description{
Retourne un objet avec uniquement ses métadonnées propre
}
\examples{
\dontrun{
extract_id(dataset)
}
}
