% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_by_column}
\alias{find_by_column}
\title{cherche dans un tibble}
\usage{
find_by_column(data, string, col, return = c("id"))
}
\arguments{
\item{data}{le dataframe/tibble dans lequel rechercher}

\item{string}{la chaine à chercher}

\item{col}{la colonne dans laquelle chercher}

\item{return_col}{a vecteur de colonnes à retourner}
}
\value{
un id
}
\description{
cherche dans un tibble
}
\examples{
\dontrun{
find_by_column(data, "un titre", "title")
find_by_column(data = data, string = "un titre", "title", c("id", "rid"))
find_by_column(data = data, string = stringr::fixed("rid"), "rid")
}
}
\keyword{internal}
