% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_alerts}
\alias{get_alerts}
\title{Récupère les datasets devant être mis à jour}
\usage{
get_alerts()
}
\value{
un tibble des datasets devant être mis à jour
}
\description{
Récupère les datasets devant être mis à jour
}
\examples{
get_alerts()
}
