% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attachment.R
\name{get_attachent}
\alias{get_attachent}
\title{Récupère un datafile}
\usage{
get_attachent(datafile = NULL, title = NULL, dataset = NULL)
}
\arguments{
\item{title}{le titre d'une pièce jointe}

\item{dataset}{optionnel l'identifiant du dataset ou un objet \code{dido_dataset()},
fournir cet argument évite un appel à l'API}

\item{attachment}{un rid de pièce jointe, un objet \code{dido_datafile()} ou \code{dido_job()}}
}
\value{
un objet \code{\link[=dido_datafile]{dido_datafile()}}
}
\description{
Permet de récupérer les données d'un datafile en utilisant soit son rid soit
son titre.
}
\details{
Lève une exception si la recherche ne retourne plus ou moins que 1 datafile
}
\examples{
\dontrun{
get_datafile("rid")
get_datafile(title = "title")
}
}
