% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attachment.R
\name{get_attachment}
\alias{get_attachment}
\title{Récupère les métadonnées d'un attachment}
\usage{
get_attachment(data = NULL, dataset = NULL)
}
\arguments{
\item{data}{le résultat d'une recherche par \code{list_attachments()}, un rid de
attachment, un objet \code{dido_attachment()} ou \code{dido_job()}}

\item{dataset}{optionnel, le dataset parent de l'attachement.}
}
\value{
un objet \code{dido_attachment()}
}
\description{
Permet de récupérer les données d'un attachment
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dataset <- add_or_update_dataset(
  title = "Un dataset pour les attachement",
  description = "Description des données statistiques",
  topic = "Transports",
  frequency = "unknown",
)

add_or_update_attachment(
  dataset = dataset,
  title = "title",
  description = "description",
  file_name = dido_example("attachment.txt")
)

dataset \%>\%
  list_attachments() \%>\%
  slice(1) \%>\%
  get_attachment()
}
\seealso{
Other attachment: 
\code{\link{add_attachment}()},
\code{\link{add_or_update_attachment}()},
\code{\link{delete_attachment}()},
\code{\link{dido_attachment}()},
\code{\link{list_attachments}()},
\code{\link{replace_attachment}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
