% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attachments.R
\name{get_attachments}
\alias{get_attachments}
\title{Liste les attachments}
\usage{
get_attachments(dataset = NULL)
}
\arguments{
\item{dataset}{optionnel, un objet dataset}
}
\value{
un tibble avec les attachments
}
\description{
Retourne la liste des attachments ou si un objet dataset est passé en argument,
la liste des attachments de ce dataset.
}
\examples{
get_attachments()
}
\seealso{
Other attachment: 
\code{\link{get_attachment}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
