% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datafile.R
\name{get_datafile}
\alias{get_datafile}
\title{Récupère un datafile}
\usage{
get_datafile(data = NULL, dataset = NULL)
}
\arguments{
\item{data}{le résultat d'une recherche par \code{list_datafiles()}, un rid de
datafile, un objet \code{dido_job()} ou \code{dido_datafile()}}

\item{dataset}{optionnel l'identifiant du dataset ou un objet
\code{dido_dataset()}}
}
\value{
un objet \code{\link[=dido_datafile]{dido_datafile()}}
}
\description{
Permet de récupérer les données d'un datafile en utilisant soit son rid soit
son titre.
}
\details{
Lève une exception si la recherche ne retourne plus ou moins que 1 datafile
}
\examples{
library(dplyr, warn.conflicts=FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test") \%>\%
  get_datafile(datafile)
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
