% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R, R/get-datafiles.R
\name{get_datafiles}
\alias{get_datafiles}
\title{Liste les datafiles}
\usage{
get_datafiles(dataset = NULL)

get_datafiles(dataset = NULL)
}
\arguments{
\item{dataset}{optionnel, un objet dataset}
}
\value{
un tibble avec les datafiles

un tibble avec les datafiles
}
\description{
Retourne la liste des datafiles ou si un objet dataset est passé en argument,
la liste des datafiles de ce dataset.
}
\examples{
get_datafiles()
get_datafiles()
}
