% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Récupère les métadonnées d'un dataset}
\usage{
get_dataset(data)
}
\arguments{
\item{data}{l'identifiant d'un dataset ou un objet \code{dido_dataset()},
\code{dido_job()} ou \code{dido_datafile()}}
}
\value{
un objet \code{dido_dataset()}
}
\description{
Permet de récupérer les métadonnées d'un dataset en utilisant soit son titre
soit un objet \code{dido_datafile()}, \code{dido_dataset()}, \code{dido_job()}
}
\details{
Lève une exception si le titre ne retourne pas un unique dataset.
}
\examples{
library(dplyr, warn.conflicts=FALSE)

ds <- list_datasets() \%>\%
  filter(title == "Données de consommation fictive") \%>\%
  get_dataset()
}
\seealso{
Other dataset: 
\code{\link{add_dataset}()},
\code{\link{dido_dataset}()},
\code{\link{update_dataset}()}
}
\concept{dataset}
