% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{get_dataset_id}
\alias{get_dataset_id}
\title{Retourne le dataset id de l'objet}
\usage{
get_dataset_id(data)
}
\arguments{
\item{data}{un objet dido_dataset, dido_datafile ou dido_job}
}
\value{
le dataset id
}
\description{
Retourne le dataset id de l'objet
}
\examples{
ds <- list_datasets()[1, ]
get_dataset_id(ds)
}
\keyword{internal}
