% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_datasets}
\alias{get_datasets}
\title{Récupère la liste des datasets}
\usage{
get_datasets()
}
\value{
un tibble des datasets de l'utilisateur
}
\description{
Récupère la liste des datasets
}
\examples{
get_datasets()
}
