% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{get_id}
\alias{get_id}
\title{Retourne le dataset id de l'objet}
\usage{
get_id(data)
}
\arguments{
\item{data}{un objet dido_dataset, dido_datafile ou dido_job}
}
\value{
le dataset id
}
\description{
Retourne le dataset id de l'objet
}
\examples{
\dontrun{
get_id(dataset)
}
}
