% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{get_job_id}
\alias{get_job_id}
\title{Retourne l'attachment id de l'objet}
\usage{
get_job_id(data)
}
\arguments{
\item{data}{un objet dido_attachment ou une chaine}
}
\value{
le rid de l'attachment
}
\description{
Retourne l'attachment id de l'objet
}
\examples{
list_jobs() \%>\% slice(1) \%>\% get_job_id()
}
\keyword{internal}
