% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_jobs}
\alias{get_jobs}
\title{Récupère tous les jobs liés à un utilisateur}
\usage{
get_jobs()
}
\value{
un tibble de jobs
}
\description{
Récupère tous les jobs liés à un utilisateur
}
\examples{
jobs <- get_jobs()
}
