% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/millesimes.R
\name{get_millesimes}
\alias{get_millesimes}
\title{Liste les millesimes}
\usage{
get_millesimes(datafile = NULL)
}
\arguments{
\item{datafile}{un objet datafile, optionnel}
}
\value{
un tibble avec les millesimes du datafile
}
\description{
Liste les millesimes
}
\examples{
get_millesimes()
}
