% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_work_env}
\alias{get_work_env}
\title{Récupère l'environnement utilisé}
\usage{
get_work_env(quiet = FALSE)
}
\arguments{
\item{quiet}{Si \code{FALSE} quand TRUE supprime les messages d'information, (\code{FALSE} par défaut)}
}
\value{
une chaine de caractère avec l'environnement utilisé. Exemple "ECOLE"
}
\description{
Si aucun environnement n'a été fixé par \code{set_work_env()}, \code{get_work_env()}
retournera le premier environnement configuré en suivant l'ordre : DEV,
PREPROD, ECOLE. Il ne fixera \strong{jamais} PROD implicitement.
}
\examples{
\dontrun{
get_work_env()
}
}
