% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-alerts.R
\name{list_alerts}
\alias{list_alerts}
\title{Récupère les datasets devant être mis à jour}
\usage{
list_alerts()
}
\value{
un tibble des datasets devant être mis à jour
}
\description{
Récupère les datasets devant être mis à jour
}
\examples{
get_alerts()
}
