% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-attachments.R
\name{list_attachments}
\alias{list_attachments}
\title{Liste les attachments}
\usage{
list_attachments(dataset = NULL)
}
\arguments{
\item{dataset}{optionnel, un objet dataset}
}
\value{
un tibble avec les attachments
}
\description{
Retourne la liste des attachments ou si un objet dataset est passé en argument,
la liste des attachments de ce dataset.
}
\examples{
# toutes les pièces jointes
list_attachments()

library(dplyr, warn.conflicts=FALSE)

# les pièces jointes d'un dataset en particulier
list_datasets() \%>\%
  filter(title == "Un jeu de données de test") \%>\%
  list_attachments()
}
\seealso{
Other attachment: 
\code{\link{add_attachment}()},
\code{\link{dido_attachment}()},
\code{\link{get_attachment}()},
\code{\link{replace_attachment}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
