% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Récupère la liste des datasets}
\usage{
list_datasets()
}
\value{
un tibble des datasets de l'utilisateur
}
\description{
Récupère la liste des datasets
}
\examples{
list_datasets()
}
