% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{list_envs}
\alias{list_envs}
\title{Liste les environnements configurés}
\usage{
list_envs()
}
\description{
Liste les environnements configurés avec les URL associées.
}
\examples{
\dontrun{
list_envs()
}
}
