% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{list_jobs}
\alias{list_jobs}
\title{Récupère tous les jobs liés à un utilisateur}
\usage{
list_jobs()
}
\value{
un tibble de jobs
}
\description{
Récupère tous les jobs liés à un utilisateur
}
\examples{
jobs <- list_jobs()
}
