% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-millesimes.R
\name{list_millesimes}
\alias{list_millesimes}
\title{Liste les millesimes}
\usage{
list_millesimes(data = NULL)
}
\arguments{
\item{data}{le résultat d'une recherche par \code{list_datafiles()}, un rid de
datafile, un objet \code{dido_job()} ou \code{dido_datafile()}}
}
\value{
un tibble avec les millesimes du datafile
}
\description{
Retourne un tibble des millésimes.
}
\examples{
library(dplyr, warn.conflicts=FALSE)

ds <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test") \%>\%
  list_millesimes()
}
\seealso{
Other millesime: 
\code{\link{add_millesime}()},
\code{\link{replace_millesime}()}
}
\concept{millesime}
