% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{load_envs}
\alias{load_envs}
\title{Recharge les environnements}
\usage{
load_envs()
}
\description{
Cette commande permet de recharger les environnements si vous avez modifié le Renviron.
Vous devez d'abord le relire avec \code{\link[=readRenviron]{readRenviron()}}
}
\examples{
\dontrun{
readEnviron()
load_envs()
}
}
