% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{me}
\alias{me}
\title{Récupère les informations de l'utilisateur}
\usage{
me()
}
\value{
Les informations de l'utilisateur
}
\description{
Récupère les informations de l'utilisateur
}
\examples{
user <- me()
}
\keyword{internal}
