% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-organization.R
\name{my_organization}
\alias{my_organization}
\title{Récupère l'ID de l'organisation de l'utilisateur.}
\usage{
my_organization(name = NULL)
}
\arguments{
\item{name}{name or acronym for organization, default to NULL}
}
\value{
un id d'organisation
}
\description{
Si l'utilisateur appartient à plusieurs organisations
}
\examples{
organization_id <- my_organization()
organization_id <- my_organization(name = "BSI")
}
