% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{organization}
\alias{organization}
\title{Get user's organization}
\usage{
organization(name = NULL)
}
\arguments{
\item{name}{name or acronym for organization, default to NULL}
}
\value{
an organization Id
}
\description{
Return the user's organization. If user belong to multiple organization, you
can specify a name.
}
\examples{
organization.id <- organization()
}
