% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-didoscalim.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{locale}
\alias{default_locale}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{readr}{\code{\link[readr:locale]{default_locale}}, \code{\link[readr]{locale}}}
}}

