% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-attachment.R
\name{replace_attachment}
\alias{replace_attachment}
\title{Remplace un fichier annexe}
\usage{
replace_attachment(attachment, file_name = NULL, remote_url = NULL)
}
\arguments{
\item{attachment}{un objet \code{dido_attachment()} ou un rid}

\item{file_name, remote_url}{le nom du fichier à verser ou l'url de l'attachement.}
}
\value{
un objet \code{dido_attachment()}
}
\description{
Remplace un fichier annexe
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dataset <- list_datasets() \%>\%
  filter(title == "Un jeu de données de test")

attachment <- list_attachments(dataset) \%>\%
  filter(title == "Un autre fichier annexe")

replace_attachment(
  attachment = attachment,
  file_name = dido_example("attachment.txt")
)
}
\seealso{
Other attachment: 
\code{\link{add_attachment}()},
\code{\link{add_or_update_attachment}()},
\code{\link{delete_attachment}()},
\code{\link{dido_attachment}()},
\code{\link{get_attachment}()},
\code{\link{list_attachments}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
