% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-millesime.R
\name{replace_millesime}
\alias{replace_millesime}
\title{Remplacer les données d'un millésime}
\usage{
replace_millesime(
  datafile,
  file_name,
  millesime,
  date_diffusion = NULL,
  quiet = NULL
)
}
\arguments{
\item{datafile}{un objet dido_datafile obtenu par \code{get_datafile()}}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}

\item{date_diffusion}{(optionnel) la date/heure à laquelle le fichier
devra être rendu accessible à la diffusion au format "AAAA-MM-JJ HH:MM:SS".
Si la time zone n'est pas précisée, l'heure locale est utilisée.

Si cette date est dans le passé, les données sont immédiatement
accessibles, si elle est dans le futur, les données ne seront accessibles
qu'à cette date/heure.

Si non précisée prend la date/heure courante (heure locale), les données
sont donc immédiatement accessibles.}

\item{quiet}{quand TRUE ou que l'option dido_quiet est à TRUE supprime les
messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{dido_job()}
}
\description{
Remplacer les données d'un millésime
}
\examples{
library(dplyr, warn.conflicts = FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test")

millesime <- replace_millesime(
  datafile = datafile,
  file_name = dido_example("augmente.csv"),
  millesime = "2021-12"
)
}
\seealso{
Other millesime: 
\code{\link{add_millesime}()},
\code{\link{list_millesimes}()}
}
\concept{millesime}
