% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{set_work_env}
\alias{set_work_env}
\title{Fixe l'environnement DiDo à utiliser}
\usage{
set_work_env(env_name = NULL, quiet = FALSE)
}
\arguments{
\item{env_name}{le nom de l'environnement à utiliser parmi PROD, ECOLE,
PREPROD, DEV et suivant ce que vous avez configuré. Si env_name n'est pas
passé, \code{set_work_env()} choisira le premier environnement configuré parmi
DEV, PREPROD et ECOLE mais \strong{jamais} PROD.}

\item{quiet}{Si \code{FALSE} quand TRUE supprime les messages d'information, (\code{FALSE} par défaut)}
}
\description{
Fixe l'environnement DiDo à utiliser
}
\examples{
\dontrun{
set_work_env("PROD")
}
}
