% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-datafile.R
\name{update_datafile}
\alias{update_datafile}
\title{Mettre à jour un les métadonnées d'un datafile.}
\usage{
update_datafile(datafile)
}
\arguments{
\item{datafile}{un objet datafile retourné par \code{get_datafile()} modifié par
l'utilisateur}
}
\value{
un objet \code{dido_job()} invisible
}
\description{
Mettre à jour un les métadonnées d'un datafile.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test") \%>\%
  get_datafile() \%>\%
  clean_metadata()

datafile$temporal_coverage$start <- "2022-01-01"
datafile$temporal_coverage$end <- "2023-12-31"

update_datafile(datafile)
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{add_or_update_datafile}()},
\code{\link{delete_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()}
}
\concept{datafile}
