% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{Met à jour les metadonnées d'un dataset}
\usage{
update_dataset(dataset)
}
\arguments{
\item{dataset}{un objet dido_dataset modifié}
}
\value{
un objet \code{dido_dataset()}
}
\description{
Cette fonction permet de mettre à jour les métadonnées d'un dataset comme,
par exemple, la date de fin de couverture temporelle.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

dataset <- list_datasets() \%>\%
  filter(title == "Un jeu de données de test") \%>\%
  get_dataset() \%>\%
  clean_metadata()

dataset$temporal_coverage$start <- "2020-01-01"
dataset$temporal_coverage$end <- "2020-12-31"
update_dataset(dataset)

update_dataset(dataset)
}
\seealso{
Other dataset: 
\code{\link{add_dataset}()},
\code{\link{dido_dataset}()},
\code{\link{get_dataset}()}
}
\concept{dataset}
