% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{upload_file}
\alias{upload_file}
\title{Verse un fichier de données à intégrer}
\usage{
upload_file(file_name)
}
\arguments{
\item{file_name}{nom du fichier}
}
\value{
un id de token
}
\description{
Verse un fichier de données à intégrer
}
\examples{
\dontrun{
token <- upload_file("mon_fichier.csv")
}
}
