% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-attachments.R
\name{add_attachment}
\alias{add_attachment}
\title{Ajoute un fichier annexe à un dataset}
\usage{
add_attachment(
  dataset,
  title,
  description,
  file_name,
  published = NULL,
  quiet = FALSE
)
}
\arguments{
\item{dataset}{l'id d'un dataset de rattachement, un objet \code{dido_dataset()},
\code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre du fichier annexe}

\item{description}{la description du fichier annexe}

\item{file_name}{le nom du fichier à verser}

\item{published}{la date de publication, si non précisée, prendra la date du jour}

\item{quiet}{quand TRUE supprime les messages d'information, \code{FALSE} par défaut}
}
\value{
un objet décrivant la pièce annexe
}
\description{
Ajoute un fichier annexe à un dataset
}
\examples{
\dontrun{
add_attachment(
  dataset = "id",
  title = "title",
  description = "description",
  file_name = "filename"
)
}
}
