% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-datafile.R
\name{add_datafile}
\alias{add_datafile}
\title{Ajoute un datafile à un dataset}
\usage{
add_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = format(Sys.time(), "\%Y-\%m"),
  published = format(Sys.time(), "\%Y-\%m-\%d"),
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S"),
  quiet = NULL
)
}
\arguments{
\item{dataset}{l'id d'un dataset, un objet \code{dido_dataset()},
\code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}

\item{published}{la date de publication du fichier, si non précisée, prend la
date du jour.}

\item{temporal_coverage_start}{optionnel, la date de début de couverture du
fichier de données au format AAAA-MM-JJ}

\item{temporal_coverage_end}{optionnel, la date de fin de couverture du
fichier de données au format AAAA-MM-JJ}

\item{legal_notice}{les mentions légales, par défaut "SDES"}

\item{date_diffusion}{(optionnel) la date/heure à laquelle le fichier
devra être rendu accessible à la diffusion au format "AAAA-MM-JJ HH:MM:SS".
Si la time zone n'est pas précisée, l'heure locale est utilisée.

Si cette date est dans le passé, les données sont immédiatement
accessibles, si elle est dans le futur, les données ne seront accessibles
qu'à cette date/heure.

Si non précisée prend la date/heure courante (heure locale), les données
sont donc immédiatement accessibles.}

\item{quiet}{quand \code{TRUE} ou que l'option dido_quiet est à \code{TRUE} supprime
les messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{dido_job()}
}
\description{
Ajoute un datafile à un dataset
}
\examples{
library(dplyr, warn.conflicts=FALSE)

dataset <- list_datasets() \%>\%
  filter(title == "Données de consommation fictive")

add_datafile(
  dataset = dataset,
  title = "titre",
  description = "description",
  file_name = dido_example("augmente.csv")
)

# publier un fichier de données avec toutes les métadonnées et un embargo
add_datafile(
  dataset = dataset,
  title = "titre 2",
  description = "description 2",
  file_name = dido_example("augmente.csv"),
  temporal_coverage_start = "2021-01-01",
  temporal_coverage_end = "2021-12-31",
  legal_notice = "something",
  millesime = "2020-10",
  date_diffusion = "2020-11-01 08:00:00"
)
}
\seealso{
Other datafile: 
\code{\link{dido_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
