% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-datafile.R
\name{add_datafile}
\alias{add_datafile}
\title{Ajoute un datafile à un dataset}
\usage{
add_datafile(
  dataset,
  title,
  description,
  file_name,
  millesime = NULL,
  published = NULL,
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = NULL,
  quiet = NULL
)
}
\arguments{
\item{dataset}{l'id d'un dataset, un objet \code{dido_dataset()},
\code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (YYYY-MM). Par défaut prendre la valeur YYYY-MM
avec l'année courante et le mois courant}

\item{published}{la date de publication du fichier, si non précisée, prend la
date du jour.}

\item{temporal_coverage_start}{la date de début des données au format
YYYY-MM-DD}

\item{temporal_coverage_end}{la date de fin des données au format YYYY-MM-DD}

\item{legal_notice}{mention légale, par défaut "SDES"}

\item{date_diffusion}{la date/heure à laquelle le fichier sera accessible au
format ISO 8601 (2021-10-01T08:00:00Z). Si non précisée prend la date du
jour à minuit, le fichier est donc immédiatement accessible.}

\item{quiet}{quand TRUE ou que l'option dido_quiet est à TRUE supprime les
messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{\link[=dido_job]{dido_job()}}
}
\description{
Ajoute un datafile à un dataset
}
\examples{
\dontrun{
datafile <- add_datafile(
  dataset = dataset$id,
  title = "titre",
  description = "description",
  token_file = file_id
)
}
}
\seealso{
Other datafile: 
\code{\link{get_datafiles}()},
\code{\link{get_datafile}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
