% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-dataset.R
\name{add_dataset}
\alias{add_dataset}
\title{Créé un dataset}
\usage{
add_dataset(
  title,
  description,
  topic,
  frequency,
  tags = NULL,
  frequency_date = NULL,
  granularity = NULL,
  zones = NULL,
  organization = NULL,
  license = "fr-lo",
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  caution = NULL,
  quiet = NULL
)
}
\arguments{
\item{quiet}{quand TRUE ou que l'option dido_quiet est à TRUE supprime les
messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{\link[=dido_dataset]{dido_dataset()}}. Ce dernier sert pour créer par la suite
les datafiles et les millésimes.
}
\description{
Créé un dataset
}
\examples{
dataset <- add_dataset(
  title = "le titre du dataset",
  description = "la description du dataset",
  topic = "Transports",
  frequency = "unknown"
)
}
