% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-millesime.R
\name{add_millesime}
\alias{add_millesime}
\title{Ajoute un millésime à un datafile}
\usage{
add_millesime(
  datafile,
  file_name,
  date_diffusion = NULL,
  millesime = NULL,
  quiet = NULL
)
}
\arguments{
\item{datafile}{un objet dido_datafile obtenu par \code{get_datafile()}}

\item{file_name}{le fichier à charger}

\item{date_diffusion}{la date/heure de diffusion au format ISO 8601, si non
précisée, le défaut est minuit du jour courant}

\item{millesime}{l'identifiant du millésime à publier, si non précisé AAAA-MM avec l'année et le mois courant}

\item{quiet}{quand TRUE ou que l'option dido_quiet est à TRUE supprime les
messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{dido_job()}
}
\description{
Ajoute un millésime à un datafile
}
\examples{
library(dplyr, warn.conflicts=FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test")

millesime <- add_millesime(
  datafile = datafile,
  file_name = dido_example("augmente.csv"),
  millesime = "2011-10"
)
}
\seealso{
Other millesime: 
\code{\link{list_millesimes}()},
\code{\link{replace_millesime}()}
}
\concept{millesime}
