% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-millesime.R
\name{add_millesime}
\alias{add_millesime}
\title{Ajoute un millésime}
\usage{
add_millesime(
  datafile,
  file_name,
  date_diffusion = NULL,
  millesime = NULL,
  quiet = TRUE
)
}
\arguments{
\item{datafile}{un objet dido_datafile}

\item{file_name}{le fichier à verser}

\item{date_diffusion}{la date/heure de diffusion au format ISO 8601, si non précisée, le défaut est minuit du jour courant}

\item{millesime}{l'identifiant du millésime à publier, si non précisé AAAA-MM avec l'année et le mois courant}

\item{quiet}{Si TRUE n'affiche pas les messages. Défaut à \code{FALSE}}
}
\value{
un job
}
\description{
Ajoute un millésime
}
\examples{
\dontrun{
millesime <- add_millesime(
  id = "id",
  rid = "rid",
  file_name = "csv_upload.csv",
  millesime = "2022-10"
)
}
}
