% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{clean_metadata}
\alias{clean_metadata}
\title{Prépare un objet pour l'envoyer à DiDo}
\usage{
clean_metadata(data)
}
\arguments{
\item{data}{objet dido_dataset, dido_datafile}
}
\value{
un objet du même type avec uniquement ses métadonnées propres.
}
\description{
Les objets \code{dido_dataset()} et \code{dido_datafile()} portent des informations sur
les objets enfants (fichiers annexes, datafile, millésime...). Cette fonction
permet de les supprimer pour ne garder que les métadonnées propres à l'objet.
}
\examples{
\dontrun{
clean_metadata(dataset)
}
}
\keyword{internal}
