% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete-dataset.R
\name{delete_dataset}
\alias{delete_dataset}
\title{Supprime un dataset}
\usage{
delete_dataset(dataset)
}
\arguments{
\item{dataset}{un identifiant de dataset}
}
\value{
TRUE
}
\description{
Supprime un dataset
}
\examples{
\dontrun{
list_datasets() \%>\%
  filter(str_detect(title, "didoscalim")) \%>\%
  select(id) \%>\%
  pmap(~ delete_dataset(..1))
}
}
\keyword{internal}
