% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-attachment.R
\name{dido_attachment}
\alias{dido_attachment}
\title{L'objet dido_attachment}
\description{
L'objet dido_attachment inclus les champs :
\itemize{
\item \code{rid} l'identifiant du jeu de données parent
\item \code{title}
\item \code{description}
\item \code{created_at}
\item \code{last_modified}
\item \code{published}
\item \code{url} : l'url du fichier annexe
}
}
\seealso{
Other attachment: 
\code{\link{add_attachment}()},
\code{\link{get_attachment}()},
\code{\link{list_attachments}()},
\code{\link{replace_attachment}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
