% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-csv.R
\name{dido_csv}
\alias{dido_csv}
\title{Génère les lignes d'entête du CSV augmenté}
\usage{
dido_csv(
  data,
  params = list(),
  locale = readr::default_locale(),
  cog_year = format(Sys.time(), "\%Y")
)
}
\arguments{
\item{data}{le dataframe à augmenter}

\item{params}{une liste nommée décrivant les caractéristiques des colonnes}

\item{locale}{la locale à utiliser}

\item{cog_year}{le millésime du COG utilisé si besoin. Par défaut prend l'année en cours}
}
\value{
un dataframe avec les 4 lignes de description du csv augmenté
}
\description{
Génère un dataframe avec les lignes d'entêtes du CSV augmenté comme premières
lignes.
}
\details{
Certains noms de variable sont connus par didoscalim qui génère
automatiquement le type et la description. La liste complète de ces
variables et des types/descriptions associés est :\tabular{lll}{
   nom de la variable \tab type \tab description \cr
   REGION \tab cog_region_AAAA \tab Code de la région \cr
   DEPARTEMENT \tab cog_departement_AAAA \tab Code du département \cr
   COMMUNE \tab cog_commune_AAAA \tab Code de la commune \cr
   EPCI \tab cog_epci_AAAA \tab Code de l'EPCI \cr
   IRIS \tab cog_iris_AAAA \tab Code de l'IRIS \cr
   ANNEE \tab n/a \tab Millésimes des données \cr
   MOIS \tab n/a \tab mois des données \cr
}


L'année \code{AAAA} est par défaut l'année courante, vous pouvez la modifier en
passant le paramètre \code{cog_year}
}
\examples{
\dontrun{
params <- list(
  OPERATEUR = list(description = "L'opérateur", type = "texte"),
  CONSO = list(description = "La consommation", unit = "Mwh")
)
augmente <- dido_csv(data, params = params)
}
}
\seealso{
En complément, vous pouvez lire : \href{https://cgdd.gitlab-pages.din.developpement-durable.gouv.fr/sdsed-bun/datalake/api/040-csvfile/}{la description d'un fichier csv augmenté},
\href{https://cgdd.gitlab-pages.din.developpement-durable.gouv.fr/sdsed-bun/datalake/api/210-headers/}{la liste des entêtes utilisables}
}
