% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-csv.R
\name{dido_csv}
\alias{dido_csv}
\title{Génère le CSV augmenté utilisé par DiDo}
\usage{
dido_csv(
  tbl,
  params = list(),
  locale = readr::default_locale(),
  cog_year = format(Sys.time(), "\%Y")
)
}
\arguments{
\item{tbl}{le dataframe/tablespace à augmenter}

\item{params}{une liste nommée décrivant les caractéristiques des colonnes}

\item{locale}{la locale à utiliser}

\item{cog_year}{le millésime du COG utilisé si besoin. Par défaut prend l'année en cours}
}
\value{
un tibble avec les 4 lignes de description
}
\description{
Génère le CSV augmenté utilisé par DiDo
}
\examples{
\dontrun{
params <- list(
  OPERATEUR = list(description = "une description", type = "enum"),
  CONSO = list(description = "une autre", unit = "Mwh")
)
dido_csv(my_tibble)
}
}
