% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-datafile.R
\name{dido_datafile}
\alias{dido_datafile}
\title{L'objet dido_datafile}
\usage{
dido_datafile(
  dataset,
  title,
  description,
  millesime = NULL,
  published = NULL,
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = NULL
)
}
\arguments{
\item{dataset}{l'id d'un dataset ou un objet dataset tel que retourné par
\code{get_dataset()} ou \code{create_dataset()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{millesime}{le millesime (YYYY-MM). Par défaut prendre la valeur YYYY-MM
avec l'année courante et le mois courant}

\item{published}{la date de publication du fichier, si non précisée, prend la
date du jour.}

\item{temporal_coverage_start}{la date de début des données au format
YYYY-MM-DD}

\item{temporal_coverage_end}{la date de fin des données au format YYYY-MM-DD}

\item{legal_notice}{mention légale, par défaut "SDES"}

\item{date_diffusion}{date/heure de diffusion du fichier au format ISO 8601
(2021-10-01T08:00:00Z). Si non précisée prend la date du jour à minuit.}
}
\value{
un objet job_result
}
\description{
L'objet dido_datafile inclus les champs :
The response object captures all information from a request.  It includes
fields:
\itemize{
\item \code{url} the url the request was actually sent to (after redirects)
\item \code{handle} the handle associated with the url
\item \code{status_code} the http status code
\item \code{header} a named list of headers returned by the server
\item \code{cookies} a named list of cookies returned by the server
\item \code{content} the body of the response, as raw vector. See \code{\link[=content]{content()}} for various ways to access the content.
\item \code{time} request timing information
\item \code{config} configuration for the request
}
}
\details{
For non-http(s) responses, some parts including the status and
header may not be interpretable the same way as http responses.
}
\examples{
\dontrun{
datafile <- create_datafile(
  dataset = dataset$id,
  title = "titre",
  description = "description",
  token_file = file_id
)
}
}
