% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-datafile.R
\name{dido_datafile}
\alias{dido_datafile}
\title{L'objet dido_datafile}
\usage{
dido_datafile(
  dataset,
  title,
  description,
  millesime = NULL,
  published = NULL,
  temporal_coverage_start = NULL,
  temporal_coverage_end = NULL,
  legal_notice = "SDES",
  date_diffusion = NULL
)
}
\arguments{
\item{dataset}{l'id d'un dataset, un objet \code{dido_dataset()},
\code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre du datafile}

\item{description}{la description du datafile}

\item{millesime}{le millesime (YYYY-MM). Par défaut prendre la valeur YYYY-MM
avec l'année courante et le mois courant}

\item{published}{la date de publication du fichier, si non précisée, prend la
date du jour.}

\item{temporal_coverage_start}{la date de début des données au format
YYYY-MM-DD}

\item{temporal_coverage_end}{la date de fin des données au format YYYY-MM-DD}

\item{legal_notice}{mention légale, par défaut "SDES"}

\item{date_diffusion}{date/heure de diffusion du fichier au format ISO 8601
(2021-10-01T08:00:00Z). Si non précisée prend la date du jour à minuit.}
}
\value{
un objet \verb{[dido_datafile()]}
}
\description{
L'objet dido_datafile est utilisé par \code{create_datafile()} et retourné par
\code{get_datafile()}
}
\examples{
\dontrun{
datafile <- create_datafile(
  dataset = dataset$id,
  title = "titre",
  description = "description",
  token_file = file_id
)
}
}
