% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-job.R
\name{dido_job}
\alias{dido_job}
\title{L'objet dido_job}
\description{
L'objet dido_job est retourné par les fonctions \code{add_datafile()},
\code{add_millesime()} et \code{replace_millesime()}, il peut être utilisé en argument
pour les fonctions \code{add_dataset()}, \code{add_datafile()}
}
