% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dido_upload_file}
\alias{dido_upload_file}
\title{Verse un fichier de données à intégrer}
\usage{
dido_upload_file(file_name)
}
\arguments{
\item{file_name}{le nom du fichier}
}
\value{
un token
}
\description{
Verse un fichier de données à intégrer
}
\examples{
\dontrun{
token <- dido_upload_file("mon_fichier.csv")
}
}
\keyword{internal}
