% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-csv.R
\name{dido_write_csv}
\alias{dido_write_csv}
\title{Enregistre le fichier CSV augmenté utilisé par DiDo.}
\usage{
dido_write_csv(tbl, file)
}
\arguments{
\item{tbl}{un tibble retourné par \code{dido_csv()}}

\item{file}{le nom du fichier}
}
\description{
Enregistre le fichier CSV augmenté utilisé par DiDo.
}
\examples{
\dontrun{
write_dido_csv(my_tibble, "fichier.csv")
}
}
