% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didoscalim.R
\docType{package}
\name{didoscalim-package}
\alias{didoscalim-package}
\alias{_PACKAGE}
\title{didoscalim: Charger des données dans DiDo}
\description{
Ce package permet de charger des données dans DiDo en passant par l'API d'alimentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mtes-mct.github.io/didoscalim/}
  \item \url{https://github.com/MTES-MCT/didoscalim}
  \item Report bugs at \url{https://github.com/MTES-MCT/didoscalim/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Chuche \email{nicolas.chuche@developpement-durable.gouv.fr}

}
\keyword{internal}
