% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attachment.R
\name{get_attachment}
\alias{get_attachment}
\title{Récupère un attachment}
\usage{
get_attachment(attachment = NULL, title = NULL, dataset)
}
\arguments{
\item{attachment}{un rid de attachment, un objet \code{dido_attachment()} ou \code{dido_job()}}

\item{title}{le titre d'un attachment}

\item{dataset}{l'identifiant d'un dataset ou un objet \code{dido_dataset()}}
}
\value{
un objet \code{\link[=dido_attachment]{dido_attachment()}}
}
\description{
Permet de récupérer les données d'un attachment en utilisant soit son rid soit
son titre.
}
\details{
Lève une exception si la recherche ne retourne plus ou moins que 1 attachment
}
\examples{
\dontrun{
get_attachment("rid", dataset = dataset)
get_attachment(title = "title", dataset = dataset)
}
}
\seealso{
Other attachment: 
\code{\link{list_attachments}()},
\code{\link{update_attachment}()}
}
\concept{attachment}
