% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datafile.R
\name{get_datafile}
\alias{get_datafile}
\title{Récupère un datafile}
\usage{
get_datafile(data = NULL, dataset = NULL)
}
\arguments{
\item{data}{le résultat d'une recherche par \code{list_datafiles()}, un des objets
\code{dido_job()} \code{dido_datafile()} oule rid d'un datafile}

\item{dataset}{optionnel l'identifiant d'un dataset ou un objet
\code{dido_dataset()}}
}
\value{
un objet \link{dido_datafile}
}
\description{
Récupère les métadonnées d'un datafile
}
\details{
Lève une exception si \code{list_datasets()} fourni en entrée retourne plus d'un
dataset.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test") \%>\%
  get_datafile(datafile)
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
