% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datafile.R
\name{get_datafile}
\alias{get_datafile}
\title{Récupère un datafile}
\usage{
get_datafile(data = NULL, title = NULL, dataset = NULL)
}
\arguments{
\item{data}{un rid de datafile, un objet \code{dido_job()} (ou un
\code{dido_datafile()} même si l'intérêt est limité).}

\item{title}{le titre d'un datafile}

\item{dataset}{optionnel l'identifiant du dataset ou un objet
\code{dido_dataset()}}
}
\value{
un objet \code{\link[=dido_datafile]{dido_datafile()}}
}
\description{
Permet de récupérer les données d'un datafile en utilisant soit son rid soit
son titre.
}
\details{
Lève une exception si la recherche ne retourne plus ou moins que 1 datafile
}
\examples{
\dontrun{
get_datafile("rid")
get_datafile(title = "title")
}
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{dido_datafile}()},
\code{\link{list_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
