% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datafile.R
\name{get_datafile}
\alias{get_datafile}
\title{Récupère un datafile}
\usage{
get_datafile(datafile = NULL, title = NULL, dataset = NULL)
}
\arguments{
\item{datafile}{un rid de datafile, un objet \code{dido_datafile()} ou \code{dido_job()}}

\item{title}{le titre d'un datafile}

\item{dataset}{optionnel l'identifiant du dataset ou un objet \code{dido_dataset()},
fournir cet argument évite un appel à l'API}
}
\value{
un objet \code{\link[=dido_datafile]{dido_datafile()}}
}
\description{
Permet de récupérer les données d'un datafile en utilisant soit son rid soit
son titre.
}
\details{
Lève une exception si la recherche ne retourne plus ou moins que 1 datafile
}
\examples{
\dontrun{
get_datafile("rid")
get_datafile(title = "title")
}
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{get_datafiles}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
