% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dido-s3.R
\name{get_datafile_rid}
\alias{get_datafile_rid}
\title{Retourne le dafile id de l'objet}
\usage{
get_datafile_rid(data)
}
\arguments{
\item{data}{un objet dido_datafile ou dido_job}
}
\value{
le datafile id
}
\description{
Retourne le dafile id de l'objet
}
\examples{
\dontrun{
get_datafile_rid(job)
}
}
\keyword{internal}
