% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-dataset.R
\name{get_dataset}
\alias{get_dataset}
\title{Récupère les métadonnées d'un dataset}
\usage{
get_dataset(data = NULL, title = NULL)
}
\arguments{
\item{data}{l'identifiant d'un dataset ou un objet \code{dido_dataset()}, \code{dido_job()} ou \code{dido_datafile()}}

\item{title}{le titre du du dataset}
}
\value{
un objet \code{\link[=dido_dataset]{dido_dataset()}}
}
\description{
Permet de récupérer les données d'un dataset en utilisant soit son titre soit
un objet \code{dido_datafile()}, \code{dido_dataset()}, \code{dido_job()}
}
\details{
Lève une exception si la recherche ne retourne pas exactement 1 dataset
}
\examples{
\dontrun{
get_dataset(id)
get_dataset(id = id)
get_dataset(title = "un titre de dataset")
}
}
