% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset}
\alias{get_dataset}
\title{Récupère les métadonnées d'un dataset}
\usage{
get_dataset(id = NULL, title = NULL)
}
\arguments{
\item{id}{l'identifiant du dataset}

\item{title}{le titre du du dataset}
}
\value{
les métadonnées du dataset
}
\description{
Permet de récupérer les données d'un dataset en utilisant soit son id soit
son titre.
}
\details{
Lève une exception si la recherche ne retourne plus ou moins que 1 dataset
trouvé est différent de 1.
}
\examples{
\dontrun{
get_dataset(id)
get_dataset(id = id)
get_dataset(title = "un titre de dataset")
}
}
