% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{get_job}
\alias{get_job}
\title{Récupére un job en particulier}
\usage{
get_job(data)
}
\arguments{
\item{data}{l'id du job}
}
\value{
un objet `dido_job()``
}
\description{
Récupére un job en particulier
}
\examples{
\dontrun{
job <- get_job("1")
}
}
