% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-datafiles.R
\name{list_datafiles}
\alias{list_datafiles}
\title{Liste les datafiles}
\usage{
list_datafiles(dataset = NULL)
}
\arguments{
\item{dataset}{optionnel, un objet dataset}
}
\value{
un tibble avec les datafiles
}
\description{
Retourne la liste des datafiles ou si un objet dataset est passé en argument,
la liste des datafiles de ce dataset.
}
\examples{
list_datafiles()
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{get_datafile}()},
\code{\link{update_datafile}()}
}
\concept{datafile}
