% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-millesimes.R
\name{list_millesimes}
\alias{list_millesimes}
\title{Liste les millesimes}
\usage{
list_millesimes(data = NULL)
}
\arguments{
\item{data}{le résultat d'une recherche par \code{list_datafiles()}, un des objets
\code{dido_job()} \code{dido_datafile()} oule rid d'un datafile}
}
\value{
un tibble avec les millesimes du datafile
}
\description{
Retourne un tibble de tous les millésimes ou d'un datafile en particulier
}
\examples{
library(dplyr, warn.conflicts = FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test")

list_millesimes(datafile)
}
\seealso{
Other millesime: 
\code{\link{add_millesime}()},
\code{\link{replace_millesime}()}
}
\concept{millesime}
