% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{organization}
\alias{organization}
\title{Récupère l'ID de l'organisation de l'utilisateur.}
\usage{
organization(name = NULL)
}
\arguments{
\item{name}{name or acronym for organization, default to NULL}
}
\value{
un id d'organisation
}
\description{
Si l'utilisateur appartient à plusieurs organisations
}
\examples{
\dontrun{
organization_id <- organization()
organization_id <- organization( name = "BSRV" )
}
}
