% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-attachment.R
\name{replace_attachment}
\alias{replace_attachment}
\title{Remplace un fichier annexe}
\usage{
replace_attachment(attachment, file_name, published = NULL, quiet = NULL)
}
\arguments{
\item{attachment}{un objet \code{dido_attachment()} ou un rid}

\item{file_name}{le nom du fichier à verser}

\item{published}{la date de publication, si non précisée, prendra la date du
jour}

\item{quiet}{quand TRUE ou que l'option dido_quiet est à TRUE supprime les
messages d'information, \code{NULL} par défaut}
}
\value{
un objet \code{dido_attachment()}
}
\description{
Remplace un fichier annexe
}
\examples{
\dontrun{
replace_attachment(
  attachment = "id",
  file_name = "filename"
)
}
}
