% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-millesime.R
\name{replace_millesime}
\alias{replace_millesime}
\title{Remplacer les données d'un millésime}
\usage{
replace_millesime(
  datafile,
  file_name,
  millesime,
  date_diffusion = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S")
)
}
\arguments{
\item{datafile}{un objet dido_datafile obtenu par \code{get_datafile()}}

\item{file_name}{le nom du fichier à charger}

\item{millesime}{le millesime (AAAA-MM). Par défaut AAAA-MM avec l'année
courante et le mois courant}

\item{date_diffusion}{(optionnel) les date et heure auxquelles le fichier
devra être rendu accessible à la diffusion.

Ce paramètre est au format "AAAA-MM-JJ HH:MM:SS" (ou ISO8601 si vous
préférez). Si la timezone n'est pas précisée, la timezone de l'ordinateur
local est utilisée.

Si cette date/heure est dans le passé, les données sont immédiatement
accessibles, si elle est dans le futur, les données ne seront accessibles
qu'à cette date/heure.

Si ce paramètre est non précisé prend la date/heure courante, les données
sont donc immédiatement accessibles.}
}
\value{
un objet \code{dido_job()}
}
\description{
Cette méthode remplace les données d'un millésime en conservant son identifiant.
}
\examples{
library(dplyr, warn.conflicts = FALSE)

datafile <- list_datafiles() \%>\%
  filter(title == "Un fichier de données de test")

millesime <- replace_millesime(
  datafile = datafile,
  file_name = dido_example("augmente.csv"),
  millesime = "2021-12"
)
}
\seealso{
Other millesime: 
\code{\link{add_millesime}()},
\code{\link{delete_millesime}()},
\code{\link{list_millesimes}()}
}
\concept{millesime}
