% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-millesime.R
\name{replace_millesime}
\alias{replace_millesime}
\title{Remplacer un millésime}
\usage{
replace_millesime(
  datafile,
  file_name,
  millesime,
  date_diffusion = NULL,
  quiet = FALSE
)
}
\arguments{
\item{datafile}{un objet dido_datafile obtenu par \code{get_datafile()}}

\item{file_name}{le fichier à charger}

\item{millesime}{l'identifiant du millésime à publier, si non précisé AAAA-MM avec l'année et le mois courant}

\item{date_diffusion}{la date/heure de diffusion au format ISO 8601, si non
précisée, le défaut est minuit du jour courant}

\item{quiet}{Si TRUE n'affiche pas les messages. Défaut à \code{FALSE}}
}
\value{
un objet \code{dido_job()}
}
\description{
Remplacer un millésime
}
\examples{
\dontrun{
datafile <- get_datafile("rid")
millesime <- replace_millesime(
  datafile = datafile,
  file_name = "csv_upload.csv",
  millesime = "2022-10"
)
}
}
