% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafiles.R, R/update-datafile.R
\name{update_datafile}
\alias{update_datafile}
\title{Mettre à jour un datafile}
\usage{
update_datafile(datafile)

update_datafile(datafile)
}
\arguments{
\item{datafile}{un objet datafile retourné par \code{get_datafile()} ou \code{extract_metadata()} modifié par l'utilisateur}

\item{id}{l'identifiant du dataset}

\item{rid}{l'identifiant du datafile}

\item{metadata}{l'objet retourné par \code{get_datafile()}. Vous pouvez le "nettoyer" avec la fonction \code{clean_datafile_metadata()}}
}
\value{


un objet \code{dido_job()}
}
\description{
Mettre à jour un datafile

Mettre à jour un datafile.
}
\examples{
\dontrun{
update_datafile("id", "rid", metadata)
}
\dontrun{
df <- get_datafile("rid")
df$temporal_coverage$end <- "2023-12-31"
update_datafile(df)
}
}
