% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-datafile.R
\name{update_datafile}
\alias{update_datafile}
\title{Mettre à jour un datafile.}
\usage{
update_datafile(datafile)
}
\arguments{
\item{datafile}{un objet datafile retourné par \code{get_datafile()} modifié par
l'utilisateur}
}
\value{
un objet \code{dido_job()}
}
\description{
Mettre à jour un datafile.
}
\examples{
\dontrun{
df <- get_datafile("rid")
df$temporal_coverage$end <- "2023-12-31"
update_datafile(df)
}
}
\seealso{
Other datafile: 
\code{\link{add_datafile}()},
\code{\link{get_datafile}()},
\code{\link{list_datafiles}()}
}
\concept{datafile}
