% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-dataset.R
\name{update_dataset}
\alias{update_dataset}
\title{Met à jour les metadonnées d'un dataset}
\usage{
update_dataset(dataset)
}
\arguments{
\item{dataset}{un objet dido_dataset modifié}
}
\value{
un objet \code{dido_dataset()}
}
\description{
Cette fonction permet de mettre à jour les métadonnées d'un dataset comme,
par exemple, la date de fin de couverture temporelle.
}
\examples{
\dontrun{
dataset <- get_dataset("id") \%>\% clean_metadata()
dataset$temporal_coverage$end <- "2022-12-31"
update_dataset(dataset)
}
}
\seealso{
Other dataset: 
\code{\link{add_dataset}()},
\code{\link{dido_dataset}()},
\code{\link{get_dataset}()}
}
\concept{dataset}
